Fix for odd org-publish (?) behaviour
Weird thing - eval-ing the top-level defun in a buffer containing init.el
resulted in the buffer being marked as modified, without any modifications. Undo
doesn't do anything except tell me there are no more changes to undo, but I'm
prompted to deal with unsaved buffers when I commit/quit.

I theorized that org publih might be using the buffer as a workspace, and
cleaning it up poorly. Whether or not this is true, wrapping the tangler and
org-publish in a temp buffer stops it from happening.